document/DRF配套资料/第2天/2_代码/day02/booktest/2_DRF魅力展示/urls.py
from booktest import views

urlpatterns = [
]

#1,url部分
from rest_framework.routers import DefaultRouter
router = DefaultRouter()
router.register('books', views.BookViewSet, base_name="books")
urlpatterns += router.urls