document/DRF配套资料/第2天/2_代码/day02/requirements.txt
amqp==2.3.2
appnope==0.1.0
asn1crypto==0.24.0
backcall==0.1.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
decorator==4.3.0
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.11.11
django-ckeditor==5.6.1
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-filter==2.0.0
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.2.0
django-js-asset==1.2.1
django-redis==4.10.0
django-reversion==3.0.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drf-extensions==0.4.0
drf-haystack==1.8.4
elasticsearch==2.4.1
et-xmlfile==1.0.1
fdfs-client-py==1.2.6
future==0.17.1
git-review==1.0
httplib2==0.9.2
idna==2.7
ipython==7.0.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4
jedi==0.13.1
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.1.1
mutagen==1.42.0
odfpy==1.4.0
olefile==0.46
openpyxl==2.6.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
pika==0.13.1
Pillow==5.3.0
prompt-toolkit==2.0.6
ptyprocess==0.6.0
pycparser==2.19
pycryptodomex==3.7.2
Pygments==2.2.0
PyJWT==1.6.4
PyMySQL==0.9.2
pyparsing==2.3.0
python-alipay-sdk==1.10.0
python-dateutil==2.8.0
pytz==2018.5
PyYAML==3.13
QQLoginTool==0.3.0
redis==2.10.6
requests==2.20.0
simplegeneric==0.8.1
six==1.11.0
svgwrite==1.2.0
tablib==0.12.1
traitlets==4.3.2
Tree==0.2.4
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
uWSGI==2.0.18
vine==1.1.4
wcwidth==0.1.7
xadmin==0.6.1
xlrd==1.2.0
xlwt==1.3.0
